argus-workers/requirements.txt
# Argus Workers - Python Dependencies

# Task Queue
celery[redis]==5.4.0
redis==5.2.1

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# HTTP Clients
httpx==0.28.1
requests==2.32.3

# AI/LLM
openai==1.59.5
anthropic==0.42.0

# Security Tools Integration
python-nmap==0.7.1

# Data Processing
pydantic==2.10.4
python-dotenv==1.0.1

# Templating (Compliance Reports)
jinja2==3.1.5

# Logging and Monitoring
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-celery==1.1.2
